switch to pcap for packet sniffing, filter by bpf on kernel side, copy less bytes per packet, custom header parser

+++ knock-cli/src/config/config.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 pub struct Rule {
     pub name: String,
     pub host: Option<String>,
-    pub sequence: Vec<i32>,
+    pub sequence: Vec<u16>,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
