+++ knockd/src/sequence/mod.rs
@@ -4,6 +4,6 @@ mod port_sequence;
 
 pub trait SequenceDetector {
     fn start(&mut self);
-    fn add_sequence(&mut self, client_ip: String, sequence: i32);
+    fn add_sequence(&mut self, client_ip: String, sequence: u16);
     fn match_sequence(&mut self, client_ip: &str) -> bool;
 }
